arm: bcm283x: Write ACPI tables
Allocate a 64KiB buffer for ACPI tables and provide SoC specific tables
for BCM2711:
- MADT
- FADT
- PPTT
- GTDT
- SPCR

Board specific tables like DSDT and SSDT are added in a separate patch.

When ACPI is enabled for a different SoC compliation will fail by
design, indicating the required functions that needs to be implemented.
When ACPI is not enabled the added code does nothing, keeping existing
behaviour.

TEST: Booted on RPi4 with only ACPI enabled, providing no FDT to the OS.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Cc: Simon Glass <sjg@chromium.org>
Cc: Matthias Brugger <mbrugger@suse.com>
Cc: Peter Robinson <pbrobinson@gmail.com>
Cc: Tom Rini <trini@konsulko.com>